Fix platform theme Android tests hanging
