+++ composeunstyled-platformtheme/src/jvmTest/kotlin/com/composeunstyled/platformtheme/PlatformThemeJvmTest.kt
@@ -1,0 +1,56 @@
+/*
+ * Copyright (c) 2026 Composable Horizons
+ *
+ * Permission is hereby granted, free of charge, to any person obtaining a copy
+ * of this software and associated documentation files (the "Software"), to deal
+ * in the Software without restriction, including without limitation the rights
+ * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+ * copies of the Software, and to permit persons to whom the Software is
+ * furnished to do so, subject to the following conditions:
+ *
+ * The above copyright notice and this permission notice shall be included in all
+ * copies or substantial portions of the Software.
+ *
+ * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+ * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+ * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+ * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+ * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+ * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
+ * SOFTWARE.
+ */
+package com.composeunstyled.platformtheme
+
+import androidx.compose.foundation.Indication
+import androidx.compose.runtime.SideEffect
+import androidx.compose.runtime.getValue
+import androidx.compose.runtime.mutableStateOf
+import androidx.compose.runtime.setValue
+import androidx.compose.ui.graphics.Color
+import androidx.compose.ui.test.runComposeUiTest
+import assertk.assertThat
+import assertk.assertions.isEqualTo
+import assertk.assertions.isFalse
+import kotlin.test.Test
+
+class PlatformThemeJvmTest {
+
+  @Test
+  fun platformIndicationUpdatesWhenColorChanges() = runComposeUiTest {
+    var color by mutableStateOf(Color.Black)
+    val indications = mutableListOf<Indication>()
+
+    setContent {
+      val indication = platformIndication(color)
+      SideEffect {
+        indications += indication
+      }
+    }
+
+    color = Color.White
+    waitForIdle()
+
+    assertThat(indications.size).isEqualTo(2)
+    assertThat(indications[0] === indications[1]).isFalse()
+  }
+}
